analyse/requirements.txt
# AIv3 System Requirements

# Core Data Processing
pandas>=1.5.0
numpy>=1.23.0
pyarrow>=10.0.0  # Required for parquet file support

# Google Cloud Storage
google-cloud-storage>=2.10.0
google-auth>=2.20.0

# Visualization
plotly>=5.14.0
matplotlib>=3.6.0

# PDF Generation (Optional - for report generation)
reportlab>=4.0.0
kaleido>=0.2.1  # For plotly image export

# Statistical Analysis
scipy>=1.10.0
statsmodels>=0.14.0

# Progress bars and utilities
tqdm>=4.65.0

# Date handling
python-dateutil>=2.8.2

# Optional ML dependencies (if using ML features)
# scikit-learn>=1.3.0
# xgboost>=1.7.0
# optuna>=3.3.0